337: Pr comments: Move some components into surf folders
