+++ packages/orchestrator-ui-components/src/components/WfoForms/formFields/types.ts
@@ -35,37 +35,7 @@ export interface ContactPerson {
     email: string;
 }
 
-export interface ImsNode {
-    id: number;
-    name: string;
-    status: string;
-}
-
 export interface Option<Value = string> {
     value: Value;
     label: string;
-}
-
-export interface ServicePort {
-    subscription_id?: string;
-    vlan?: string;
-    bandwidth?: number;
-    nonremovable?: boolean;
-    modifiable?: boolean;
-}
-
-export enum PortMode {
-    TAGGED = 'tagged',
-    UNTAGGED = 'untagged',
-    LINK_MEMBER = 'link_member',
-}
-
-export enum ProductTag {
-    MSC = 'MSC',
-    MSCNL = 'MSCNL',
-    IRBSP = 'IRBSP',
-    SP = 'SP',
-    SPNL = 'SPNL',
-    AGGSP = 'AGGSP',
-    AGGSPNL = 'AGGSPNL',
 }
