+++ packages/orchestrator-ui-components/src/components/WfoForms/formFields/utils.spec.ts
@@ -9,7 +9,7 @@ import {
     subscriptionHasTaggedPortModeInstanceValue,
     subscriptionHasTaggedProduct,
 } from './utils';
-import { ProductTag } from './types';
+import { ProductTag } from './surf/types';
 
 const testProductBlockInstances: ProductBlockInstance[] = [
     {
